requirements.txt
fastapi~=0.115.0
uvicorn
strawberry-graphql
pymongo~=4.10.1
